refactor: OrderService 내 Order 생성 로직 최적화 및 중복 코드 정리

+++ backend/project1/src/main/java/com/team4/project1/domain/item/controller/ApiV1ItemController.java
@@ -5,7 +5,6 @@ import com.team4.project1.domain.item.service.ItemService;
 import com.team4.project1.global.dto.ResponseDto;
 import com.team4.project1.global.exception.ItemNotFoundException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,7 +18,7 @@ public class ApiV1ItemController {
     private final ItemService itemService;
 
     @GetMapping("/{itemId}")
-    public ResponseEntity<ResponseDto<ItemDto>> item(@PathVariable Integer itemId) {
+    public ResponseEntity<ResponseDto<ItemDto>> item(@PathVariable Long itemId) {
         return ResponseEntity.ok(ResponseDto.ok(
                 itemService.getItemById(itemId).orElseThrow(() -> new ItemNotFoundException(itemId)
                 )
